requirements.txt
streamlit
fastapi
tensorflow-macos
uvicorn
python-multipart
